change class in different file
